requirements.txt
discord.py
requests
python-dotenv
cdp-sdk
coinbase